Support calling "mirage stop" without specifiying a port if only one mirage instance is running

+++ lib/mirage/client.rb
@@ -17,6 +17,7 @@ module Mirage
     method_option :port, :aliases => "-p", :type => :numeric, :default => 7001, :desc => "port that mirage should be started on"
     method_option :defaults, :aliases => "-d", :type => :string, :default => 'responses', :desc => "location to load default responses from"
     method_option :debug, :type => :boolean, :default => false, :desc => "run in debug mode"
+
     def start
 
       unless mirage_process_ids([options[:port]]).empty?
@@ -33,8 +34,7 @@ module Mirage
       end
 
 
-
-      command = command.concat(options.to_a).flatten.collect{|arg| arg.to_s}
+      command = command.concat(options.to_a).flatten.collect { |arg| arg.to_s }
       ChildProcess.build(*command).start
 
       mirage_client = Mirage::Client.new "http://localhost:#{options[:port]}/mirage"
@@ -49,43 +49,55 @@ module Mirage
     end
 
     desc "stop", "stops mirage"
-    method_option :port, :aliases => "-p", :type => :array, :required => true, :banner => "[port_1 port_2|all]", :desc => "port(s) of mirage instance(s). ALL stops all running instances"
+    method_option :port, :aliases => "-p", :type => :array, :banner => "[port_1 port_2|all]", :desc => "port(s) of mirage instance(s). ALL stops all running instances"
+
     def stop
-      mirage_process_ids(options[:port]).each do |process_id|
+
+      ports = options[:port]
+      if ports.nil?
+        mirage_process_ids = mirage_process_ids([:all])
+        raise "Mirage is running on ports #{mirage_process_ids.keys.join(", ")}. Please run mirage stop -p [PORT(s)] instead" if mirage_process_ids.size > 1
+        ports = [:all]
+      end
+
+      mirage_process_ids(ports).values.each do |process_id|
         puts "killing #{process_id}"
         ChildProcess.windows? ? `taskkill /F /T /PID #{process_id}` : `kill -9 #{process_id}`
       end
+
       wait_until do
-        mirage_process_ids(options[:port]).size == 0
+        mirage_process_ids(ports).size == 0
       end
     end
 
     private
     def mirage_process_ids ports
-      if ports.first.to_s.downcase == "all"
-        puts "in here"
 
+      mirage_instances = {}
+      if ports.first.to_s.downcase == "all"
         if ChildProcess.windows?
           [`tasklist /V | findstr "mirage\\ server"`.split(' ')[1]].compact
         else
-          ["Mirage Server", "mirage_server"].collect do |process_name|
-            puts `ps aux | grep "#{process_name}" | grep -v grep`
-            `ps aux | grep "#{process_name}" | grep -v grep`.chomp.lines.collect{|process_line| process_line.split(' ')[1]}
+          ["Mirage Server", "mirage_server"].each do |process_name|
+            `ps aux | grep "#{process_name}" | grep -v grep`.chomp.lines.collect{|line|line.chomp}.each do |process_line|
+              pid = process_line.split(' ')[1]
+              port = process_line[/port (\d+)/,1]
+              mirage_instances[port] = pid
+            end
           end.flatten.find_all { |process_id| process_id != $$.to_s }.compact
         end
-
       else
-
         ports.collect do |port|
           begin
-            http_get("http://localhost:#{port}/mirage/pid").body.to_i
+            pid = http_get("http://localhost:#{port}/mirage/pid").body.to_i
+            mirage_instances[port] = pid
           rescue
             nil
           end
-
         end.compact
-
       end
+
+      mirage_instances
     end
 
   end
@@ -95,13 +107,13 @@ module Mirage
 
 
     def start options={:port => 7001}
-      Runner.new.invoke(:start,[],options)
+      Runner.new.invoke(:start, [], options)
     end
 
     def stop options
       options[:port] = [options[:port]] unless options[:port].is_a?(Array)
       puts "Stopping Mirage"
-      Runner.new.invoke(:stop,[],options)
+      Runner.new.invoke(:stop, [], options)
     end
 
     private
